+++ ethclient/src/main/scala/fluence/ethclient/package.scala
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2018 Fluence Labs Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fluence
+
+import java.util.concurrent.{CancellationException, CompletableFuture, CompletionException}
+
+import cats.effect.Async
+
+package object ethclient {
+
+  /**
+   * Utility converters from Java's CompletableFuture to cats-effect types
+   */
+  private[ethclient] implicit class FromJavaFuture[A](cf: CompletableFuture[A]) {
+
+    /**
+     * Convert Java Future to some Async F type.
+     * It requires F to be just Async.
+     */
+    def asAsync[F[_]: Async]: F[A] =
+      Async[F].async { cb ⇒
+        cf.handle[Unit] { (result: A, err: Throwable) ⇒
+          err match {
+            case null =>
+              cb(Right(result))
+            case _: CancellationException =>
+              ()
+            case ex: CompletionException if ex.getCause ne null =>
+              cb(Left(ex.getCause))
+            case ex =>
+              cb(Left(ex))
+          }
+        }
+      }
+  }
+
+}
